Adjust PATH to not override system binaries with user binaries.
This is in an effort to reduce the possibility of an attack where an
untrusted binary overrides a trusted system binary and intercepts input
to which it shouldn't have access. Overriding a system binary is now
done via aliases in a more explicit manner which is at least not worse.

This also included deleting the search script, which is fine because I
don't really use it.